requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
cssselect==1.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
html2text==2020.1.16
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jieba3k==0.35.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
llama-cloud==0.0.6
llama-index==0.10.50
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.50
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.1
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.24
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-index-readers-web==0.1.20
llama-parse==0.4.4
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
nltk==3.8.1
numpy==1.26.4
openai==1.35.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==10.3.0
playwright==1.44.0
Protego==0.3.1
protobuf==5.27.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==11.1.0
Pygments==2.18.0
PyHamcrest==2.1.0
pyOpenSSL==24.1.0
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
rpds-py==0.18.1
streamlit==1.36.0
striprtf==0.0.26
tenacity==8.4.2
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
Twisted==20.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
w3lib==2.2.1
watchdog==4.0.1

